Add embedded Wufoo form to modal
Added an embedded wufoo form to the modal to allow future sponsors to request to be contacted to make a contribution to the Omaha GWC. This form includes automatic notifications to omahagirlswhocode@gmail.com and conditional logic based on the sponsor type (company/individual). We request that if it is a company, they provide logo and website details up front.